# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: PawonWarga Backend API Service
  title: PawonWarga API
  version: "1.0"
paths:
...
  /mentions:
    get:
      description: Lists sentiment-labeled posts for the monitoring dashboard, with
        metrics and per-platform volume aggregated over the whole filtered set. Response
        messages are localized via ?lang= or Accept-Language (id/en).
      parameters:
      - description: Free-text search in content/author
        in: query
        name: q
        type: string
      - description: X, Instagram, TikTok, News, YouTube, or \
        in: query
        name: platform
        type: string
      - description: positive, negative, neutral, or \
        in: query
        name: sentiment
        type: string
      - description: Start date (YYYY-MM-DD, WIB) — omit for no lower bound
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD, WIB, inclusive) — omit for no upper bound
        in: query
        name: to
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: newest (default), oldest, or engagement
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List mentions
      tags:
      - mentions
  /mentions/{id}:
    get:
      description: Returns a single sentiment-labeled mention. 404s if the post exists
        but hasn't been labeled yet. Response messages are localized via ?lang= or
        Accept-Language (id/en).
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get mention detail
      tags:
      - mentions
  /mentions/{id}/comments:
    get:
      description: Paginated labeled comments on a single post, oldest first. 404s
        if the post doesn't exist or hasn't been labeled yet — same check GetByID
        applies. Response messages are localized via ?lang= or Accept-Language (id/en).
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List comments for a mention
      tags:
      - mentions
...
securityDefinitions:
  ApiKeyAuth:
    description: Shared secret for the internal ingest routes (Python labeling worker)
    in: header
    name: X-API-Key
    type: apiKey
  BasicAuth:
    type: basic
  BearerAuth:
    description: 'Enter: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"